+++ .github/workflows/run_tests_ubuntu.yml
@@ -4,7 +4,7 @@
 
 name: Run Ubuntu/Linux netCDF Tests
 
-on: [ pull_request ]
+on: [pull_request,push]
 
 jobs:
