+++ .github/workflows/run_tests_win_mingw.yml
@@ -7,7 +7,7 @@
 name: Run MSYS2, MinGW64-based Tests
 
 
-on: [pull_request]
+on: [pull_request,push]
 
 jobs:
