+++ libnetcdf.settings.in
@@ -9,6 +9,7 @@ Configured On:		@CONFIG_DATE@
 Host System:		@host_cpu@-@host_vendor@-@host_os@
 Build Directory: 	@abs_top_builddir@
 Install Prefix:         @prefix@
+Plugin Install Prefix:  @PLUGIN_INSTALL_DIR_SETTING@
 
 # Compiling Options
 -----------------
