+++ nc_test4/findplugin.in
@@ -9,12 +9,10 @@
 #   $2 is 1 if we are running under cmake
 #   $3 is 1 if we are running using Visual Studio, blank otherwise
 #   $4 is the build type; only used if $3 is 1
-# Optional Input:
-#   HDF5_PLUGIN_PATH environment variable
 # Outputs:
 #   return code is 0 is success, 1 if failed
 #   Variable HDF5_PLUGIN_LIB is set to the library file name
-#   Variable HDF5_PLUGIN_PATH is setthe absolute path to the
+#   Variable HDF5_PLUGIN_DIR is set to the absolute path to the
 #                    directory containing the plugin library file
 # Local variables are prefixed with FP_
 #
@@ -27,7 +25,7 @@ findplugin() {
 FP_NAME="$1"
 
 FP_PLUGIN_LIB=
-FP_PLUGIN_PATH=
+FP_PLUGIN_DIR=
 
 # Figure out the plugin file name
 # Test for visual studio before cygwin since both might be true
@@ -43,14 +41,6 @@ else # Presumably some form on *nix"
   FP_PLUGIN_LIB="lib${FP_NAME}.so"
 fi
 
-# If HDF5_PLUGIN_PATH is defined, then it overrides everything else.
-if test "x$HDF5_PLUGIN_PATH" != x ; then
-    HDF5_PLUGIN_LIB="$FP_PLUGIN_LIB"
-    # HDF5_PLUGIN_PATH already set
-    HDF5_PLUGIN_PATH=`${NCPATHCVT} $HDF5_PLUGIN_PATH`
-    return 0;
-fi
-
 # Figure out the path to where the lib is stored
 # This can probably be simplified
 
@@ -60,39 +50,37 @@ FP_PLUGINS=`pwd`
 cd ${CURWD}
 
 # Case 1: Cmake with Visual Studio
-# Do not know where to look for a dylib
-# Case 1: Cmake with Visual Studio
 if test "x$FP_ISCMAKE" != x -a "x${FP_ISMSVC}" != x ; then
     # Case 1a: ignore the build type directory
     if test -f "${FP_PLUGINS}/${FP_PLUGIN_LIB}" ; then
-      FP_PLUGIN_PATH="${FP_PLUGINS}"
+      FP_PLUGIN_DIR="${FP_PLUGINS}"
     fi
 else # Case 2: automake
   # Case 2a: look in .libs
   if test -f "${FP_PLUGINS}/.libs/${FP_PLUGIN_LIB}" ; then
-    FP_PLUGIN_PATH="${FP_PLUGINS}/.libs"
+    FP_PLUGIN_DIR="${FP_PLUGINS}/.libs"
   else # Case 2: look in FP_PLUGINS directly
     if test -f "${FP_PLUGINS}/${FP_PLUGIN_LIB}" ; then
-      FP_PLUGIN_PATH="${FP_PLUGINS}"
+      FP_PLUGIN_DIR="${FP_PLUGINS}"
     fi
   fi
 fi
 
 # Verify
-if test "x$FP_PLUGIN_PATH" = x ; then
-  echo "***Fail: Could not locate a usable HDF5_PLUGIN_PATH"
+if test "x$FP_PLUGIN_DIR" = x ; then
+  echo "***Fail: Could not locate a usable HDF5_PLUGIN_DIR"
   return 1
 fi
-if ! test -f "$FP_PLUGIN_PATH/$FP_PLUGIN_LIB" ; then
+if ! test -f "$FP_PLUGIN_DIR/$FP_PLUGIN_LIB" ; then
   echo "***Fail: Could not locate a usable HDF5_PLUGIN_LIB"
   return 1
 fi
 
-FP_PLUGIN_PATH=`${NCPATHCVT} $FP_PLUGIN_PATH`
+FP_PLUGIN_DIR=`${NCPATHCVT} $FP_PLUGIN_DIR`
 
 # Set the final output variables
 HDF5_PLUGIN_LIB="$FP_PLUGIN_LIB"
-HDF5_PLUGIN_PATH="$FP_PLUGIN_PATH"
+HDF5_PLUGIN_DIR="$FP_PLUGIN_DIR"
 
 return 0
 }
