+++ nc_test4/tst_filter.sh
@@ -43,36 +43,37 @@ sed -e 's/[ 	]*\([^ 	].*\)/\1/' <$1 >$2
 
 # Hide/unhide the noop filter
 hidenoop() {
-  rm -fr ${HDF5_PLUGIN_PATH}/save
-  mkdir ${HDF5_PLUGIN_PATH}/save
-  mv ${NOOPPATH} ${HDF5_PLUGIN_PATH}/save
+  rm -fr ${HDF5_PLUGIN_DIR}/save
+  mkdir ${HDF5_PLUGIN_DIR}/save
+  mv ${NOOPDIR} ${HDF5_PLUGIN_DIR}/save
 }
 
 unhidenoop() {
-  mv ${HDF5_PLUGIN_PATH}/save/${NOOPLIB} ${HDF5_PLUGIN_PATH}
-  rm -fr ${HDF5_PLUGIN_PATH}/save
+  mv ${HDF5_PLUGIN_DIR}/save/${NOOPLIB} ${HDF5_PLUGIN_DIR}
+  rm -fr ${HDF5_PLUGIN_DIR}/save
 }
 
-# Locate the plugin path and the library names; argument order is critical
+# Locate the plugin dir and the library names; argument order is critical
 # Find bzip2 and capture
 findplugin h5bzip2
 BZIP2LIB="${HDF5_PLUGIN_LIB}"
-BZIP2PATH="${HDF5_PLUGIN_PATH}/${BZIP2LIB}"
+BZIP2DIR="${HDF5_PLUGIN_DIR}/${BZIP2LIB}"
 # Find misc and capture
 findplugin h5misc
-MISCPATH="${HDF5_PLUGIN_PATH}/${HDF5_PLUGIN_LIB}"
+MISCDIR="${HDF5_PLUGIN_DIR}/${HDF5_PLUGIN_LIB}"
 # Find noop and capture
 findplugin h5noop
 NOOPLIB="${HDF5_PLUGIN_LIB}"
-NOOPPATH="${HDF5_PLUGIN_PATH}/${HDF5_PLUGIN_LIB}"
+NOOPDIR="${HDF5_PLUGIN_DIR}/${HDF5_PLUGIN_LIB}"
 
-echo "final HDF5_PLUGIN_PATH=${HDF5_PLUGIN_PATH}"
-export HDF5_PLUGIN_PATH
+echo "final HDF5_PLUGIN_DIR=${HDF5_PLUGIN_DIR}"
+export HDF5_PLUGIN_DIR
+export HDF5_PLUGIN_PATH="$HDF5_PLUGIN_DIR"
 
 # Verify
-if ! test -f ${BZIP2PATH} ; then echo "Unable to locate ${BZIP2PATH}"; exit 1; fi
-if ! test -f ${MISCPATH} ; then echo "Unable to locate ${MISCPATH}"; exit 1; fi
-if ! test -f ${NOOPPATH} ; then echo "Unable to locate ${NOOPPATH}"; exit 1; fi
+if ! test -f ${BZIP2DIR} ; then echo "Unable to locate ${BZIP2DIR}"; exit 1; fi
+if ! test -f ${MISCDIR} ; then echo "Unable to locate ${MISCDIR}"; exit 1; fi
+if ! test -f ${NOOPDIR} ; then echo "Unable to locate ${NOOPDIR}"; exit 1; fi
 
 # See if we have szip
 if avail szip; then HAVE_SZIP=1; else HAVE_SZIP=0; fi
