+++ nczarr_test/Makefile.am
@@ -154,7 +154,7 @@ EXTRA_DIST += ref_power_901_constants.zip ref_power_901_constants.cdl ref_quotes
 
 CLEANFILES = ut_*.txt ut*.cdl tmp*.nc tmp*.cdl tmp*.txt tmp*.dmp tmp*.zip tmp*.nc tmp*.dump tmp*.tmp tmp_ngc.c ref_zarr_test_data.cdl
 
-BUILT_SOURCES = test_quantize.c test_filter_avail.c run_specific_filters.sh
+BUILT_SOURCES = test_quantize.c test_filter_avail.c run_specific_filters.sh run_filterinstall.sh
 test_quantize.c: $(top_srcdir)/nc_test4/tst_quantize.c
 	rm -f $@
 	echo "#define TESTNCZARR" > $@
@@ -173,6 +173,15 @@ run_specific_filters.sh: $(top_srcdir)/nc_test4/tst_specific_filters.sh
 	tr -d '\r' < run_specific_filters.tmp > $@
 	chmod a+x $@
 	rm -f run_specific_filters.tmp
+
+run_filterinstall.sh: $(top_srcdir)/nc_test4/tst_filterinstall.sh
+	rm -f $@ run_filterinstall.tmp
+	echo "#!/bin/bash" > run_filterinstall.tmp
+	echo "TESTNCZARR=1" >> run_filterinstall.tmp
+	cat $(top_srcdir)/nc_test4/tst_filterinstall.sh >> run_filterinstall.tmp
+	tr -d '\r' < run_filterinstall.tmp > $@
+	chmod a+x $@
+	rm -f run_filterinstall.tmp
 
 # Remove directories
 clean-local:
