+++ nczarr_test/run_filter.sh
@@ -51,21 +51,21 @@ sed -e 's/[ 	]*\([^ 	].*\)/\1/' <$1 >$2
 
 # Find misc and capture
 findplugin h5misc
-MISCPATH="${HDF5_PLUGIN_PATH}/${HDF5_PLUGIN_LIB}"
+MISCDIR="${HDF5_PLUGIN_DIR}/${HDF5_PLUGIN_LIB}"
 
 # Find noop and capture
 findplugin h5noop
 NOOPLIB="${HDF5_PLUGIN_LIB}"
-NOOPPATH="${HDF5_PLUGIN_PATH}/${NOOPLIB}"
+NOOPDIR="${HDF5_PLUGIN_DIR}/${NOOPLIB}"
 
 # Find bzip2 and capture
 findplugin h5bzip2
 BZIP2LIB="${HDF5_PLUGIN_LIB}"
-BZIP2PATH="${HDF5_PLUGIN_PATH}/${BZIP2LIB}"
+BZIP2DIR="${HDF5_PLUGIN_DIR}/${BZIP2LIB}"
 
 # Verify
-if ! test -f ${BZIP2PATH} ; then echo "Unable to locate ${BZIP2PATH}"; exit 1; fi
-if ! test -f ${MISCPATH} ; then echo "Unable to locate ${MISCPATH}"; exit 1; fi
+if ! test -f ${BZIP2DIR} ; then echo "Unable to locate ${BZIP2DIR}"; exit 1; fi
+if ! test -f ${MISCDIR} ; then echo "Unable to locate ${MISCDIR}"; exit 1; fi
 
 # Execute the specified tests
 
@@ -144,11 +144,11 @@ ${NCDUMP} -hs $fileurl > ./tmp_known_$zext.txt
 # Remove irrelevant -s output
 sclean ./tmp_known_$zext.txt tmp_known_$zext.dump
 # Now hide the filter code
-mv ${NOOPPATH} ./${NOOPLIB}.save
+mv ${NOOPDIR} ./${NOOPLIB}.save
 # dump and clean noop.nc header when filter is not avail
 ${NCDUMP} -hs $fileurl > ./tmp_unk_$zext.txt
 # Restore the filter code
-mv ./${NOOPLIB}.save ${NOOPPATH}
+mv ./${NOOPLIB}.save ${NOOPDIR}
 # Verify that the filter is no longer defined
 UNK=`sed -e '/var:_Filter/p' -e d ./tmp_unk_$zext.txt`
 test "x$UNK" = x
