+++ plugins/CMakeLists.txt
@@ -10,12 +10,6 @@ FILE(READ H5Znoop.c NOOP_SOURCE)
 FILE(WRITE ${CMAKE_CURRENT_BINARY_DIR}/H5Znoop1.c "#define NOOP_INSTANCE 1\n")
 FILE(APPEND ${CMAKE_CURRENT_BINARY_DIR}/H5Znoop1.c "${NOOP_SOURCE}")
 
-SET(h5bzip2_SOURCES H5Zbzip2.c)
-
-IF(NOT HAVE_BZIP2)
-SET(h5bzip2_SOURCES ${h5bzip2_SOURCES} blocksort.c huffman.c crctable.c randtable.c compress.c decompress.c bzlib.c)
-ENDIF()
-
 SET(h5misc_SOURCES H5Zmisc.c H5Zutil.c h5misc.h)
 
 SET(h5noop_SOURCES H5Znoop.c H5Zutil.c h5noop.h)
@@ -29,11 +23,10 @@ SET(h5fletcher32_SOURCES H5Zfletcher32.c H5checksum.c)
 SET(h5deflate_SOURCES H5Zdeflate.c)
 
 SET(nczmisc_SOURCES NCZmisc.c)
-SET(nczdefaults_SOURCES NCZdefaults.c)
+SET(nczhdf5filters_SOURCES NCZhdf5filters.c)
+SET(nczstdfilters_SOURCES NCZstdfilters.c)
 
-IF(ENABLE_FILTER_TESTING)
-IF(BUILD_UTILITIES)
-
+IF(ENABLE_PLUGINS)
 # LDFLAGS = -module -avoid-version -shared -export-dynamic -no-undefined
 
 SET(CMAKE_LIBRARY_OUTPUT_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}")
@@ -64,7 +57,8 @@ buildplugin(h5fletcher32 "h5fletcher32")
 buildplugin(h5deflate "h5deflate")
 
 buildplugin(nczmisc "nczmisc")
-buildplugin(nczdefaults "nczdefaults" netcdf)
+buildplugin(nczhdf5filters "nczhdf5filters" netcdf)
+buildplugin(nczstdfilters "nczstdfilters" netcdf)
 
 IF(ENABLE_BLOSC)
 SET(h5blosc_SOURCES H5Zblosc.c)
@@ -80,17 +74,54 @@ ENDIF()
 IF(HAVE_SZ)
   SET(h5szip_SOURCES H5Zszip.c H5Zszip.h)
   buildplugin(h5szip "h5szip" ${Szip_LIBRARIES})
-  SET(nczszip_SOURCES NCZszip.c)
-  buildplugin(nczszip "nczszip" netcdf)
 ENDIF()
 
+IF(HAVE_LOCAL_BZ2)
+SET(h5bzip2_SOURCES H5Zbzip2.c blocksort.c huffman.c crctable.c randtable.c compress.c decompress.c bzlib.c bzlib.h bzlib_private.h)
+buildplugin(h5bzip2 "h5bzip2")
+ELSE()
+SET(h5bzip2_SOURCES H5Zbzip2.c)
 buildplugin(h5bzip2 "h5bzip2" ${Bzip2_LIBRARIES})
-# Note we use name h5bzip2 instead of bzip2 to avoid logical
-# target name clash with examples/C/hdf5plugins
-SET_TARGET_PROPERTIES(h5bzip2 PROPERTIES OUTPUT_NAME "bzip2")
+ENDIF()
 
-ENDIF(BUILD_UTILITIES)
-ENDIF(ENABLE_FILTER_TESTING)
+
+# Installation
+IF(ENABLE_PLUGIN_INSTALL)
+
+MACRO(installplugin PLUG)
+IF(MSVC)
+  SET(BUILD_PLUGIN_LIB "${PLUG}.dll")
+ELSE()
+   SET(BUILD_PLUGIN_LIB "lib${PLUG}.so")
+ENDIF()
+MESSAGE(STATUS "Installing: ${BUILD_PLUGIN_LIB} into ${PLUGIN_INSTALL_DIR}")
+install(PROGRAMS ${CMAKE_CURRENT_BINARY_DIR}/${BUILD_PLUGIN_LIB} DESTINATION ${PLUGIN_INSTALL_DIR})
+ENDMACRO()
+
+install(DIRECTORY DESTINATION ${PLUGIN_INSTALL_DIR})
+IF(Bzip2_FOUND)
+installplugin(h5bzip2)
+ENDIF()
+IF(Zstd_FOUND)
+installplugin(h5zstd)
+ENDIF()
+IF(Blosc_FOUND)
+installplugin(h5blosc)
+ENDIF()
+IF(ENABLE_NCZARR)
+  installplugin(h5fletcher32)
+  installplugin(h5shuffle)
+  installplugin(h5deflate)
+  installplugin(nczhdf5filters)
+  installplugin(nczstdfilters)
+IF(Szip_FOUND)
+  installplugin(h5szip)
+ENDIF()
+ENDIF()
+
+ENDIF(ENABLE_PLUGIN_INSTALL)
+
+ENDIF(ENABLE_PLUGINS)
 
 # Copy some test files from current source dir to out-of-tree build dir.
 FILE(COPY ${COPY_FILES} DESTINATION ${CMAKE_CURRENT_BINARY_DIR}/)
