+++ plugins/H5Zzstd.c
@@ -166,92 +166,3 @@ cleanupAndFail:
    return 0;
 }
 
-/**************************************************/
-/* NCZarr Filter Objects */
-
-/* Provide the codec support for the HDF5 zstandard library */
-
-static int NCZ_zstd_codec_to_hdf5(const char* codec, size_t* nparamsp, unsigned** paramsp);
-static int NCZ_zstd_hdf5_to_codec(size_t nparams, const unsigned* params, char** codecp);
-
-static NCZ_codec_t NCZ_zstd_codec = {/* NCZ_codec_t  codec fields */ 
-  NCZ_CODEC_CLASS_VER,	/* Struct version number */
-  NCZ_CODEC_HDF5,	/* Struct sort */
-  "zstd",	        /* Standard name/id of the codec */
-  H5Z_FILTER_ZSTD,   /* HDF5 alias for zstd */
-  NULL, /*NCZ_zstd_codec_initialize*/
-  NULL, /*NCZ_zstd_codec_finalize*/
-  NCZ_zstd_codec_to_hdf5,
-  NCZ_zstd_hdf5_to_codec,
-  NULL, /*NCZ_zstd_modify_parameters*/
-};
-
-/* External Export API */
-DLLEXPORT
-const void*
-NCZ_get_codec_info(void)
-{
-    return (void*)&NCZ_zstd_codec;
-}
-
-static int
-NCZ_zstd_codec_to_hdf5(const char* codec_json, size_t* nparamsp, unsigned** paramsp)
-{
-    int stat = NC_NOERR;
-    NCjson* jcodec = NULL;
-    NCjson* jtmp = NULL;
-    unsigned* params = NULL;
-    struct NCJconst jc;
-  
-    if(nparamsp == NULL || paramsp == NULL)
-        {stat = NC_EINTERNAL; goto done;}
-    
-    if((params = (unsigned*)calloc(1,sizeof(unsigned)))== NULL)
-        {stat = NC_ENOMEM; goto done;}
-
-    /* parse the JSON */
-    if(NCJparse(codec_json,0,&jcodec))
-        {stat = NC_EFILTER; goto done;}
-    if(NCJsort(jcodec) != NCJ_DICT) {stat = NC_EPLUGIN; goto done;}
-    /* Verify the codec ID */
-    if(NCJdictget(jcodec,"id",&jtmp))
-        {stat = NC_EFILTER; goto done;}
-    if(jtmp == NULL || !NCJisatomic(jtmp)) {stat = NC_EFILTER; goto done;}
-    if(strcmp(NCJstring(jtmp),NCZ_zstd_codec.codecid)!=0) {stat = NC_EINVAL; goto done;}
-
-    /* Get Level */
-    if(NCJdictget(jcodec,"level",&jtmp))
-        {stat = NC_EFILTER; goto done;}
-    if(NCJcvt(jtmp,NCJ_INT,&jc))
-        {stat = NC_EFILTER; goto done;}
-    if(jc.ival < 0 || jc.ival > NC_MAX_UINT) {stat = NC_EINVAL; goto done;}
-    params[0] = (unsigned)jc.ival;
-    *nparamsp = 1;
-    *paramsp = params; params = NULL;
-    
-done:
-    if(params) free(params);
-    NCJreclaim(jcodec);
-    return stat;
-}
-
-static int
-NCZ_zstd_hdf5_to_codec(size_t nparams, const unsigned* params, char** codecp)
-{
-    int stat = NC_NOERR;
-    unsigned level = 0;
-    char json[1024];
-
-    if(nparams == 0 || params == NULL)
-        {stat = NC_EFILTER; goto done;}
-
-    level = params[0];
-    snprintf(json,sizeof(json),"{\"id\": \"%s\", \"level\": \"%u\"}",NCZ_zstd_codec.codecid,level);
-    if(codecp) {
-        if((*codecp = strdup(json))==NULL) {stat = NC_ENOMEM; goto done;}
-    }
-    
-done:
-    return stat;
-}
-
