+++ test_common.in
@@ -28,9 +28,11 @@ FEATURE_PARALLEL=@HAS_PARALLEL@
 
 # Define selected features of the build
 FEATURE_HDF5=@HAS_HDF5@
+FEATURE_NCZARR=@HAS_NCZARR@
 FEATURE_S3TESTS=@DO_NCZARR_S3_TESTS@
 FEATURE_NCZARR_ZIP=@DO_NCZARR_ZIP_TESTS@
 FEATURE_FILTERTESTS=@DO_FILTER_TESTS@
+FEATURE_PLUGIN_INSTALL_DIR=@PLUGIN_INSTALL_DIR@
 
 set -e
